[REFACTOR] #6 - import 정보 수정
