+++ src/main/kotlin/com/example/smartshopping/presentation/controller/ProductApiController.kt
@@ -2,14 +2,14 @@ package com.example.smartshopping.presentation.controller
 
 import com.example.smartshopping.common.ApiResponse
 import com.example.smartshopping.common.SmartShoppingException
-import com.example.smartshopping.domain.product.ProductService
-import com.example.smartshopping.domain.product.registration.ProductImageService
-import com.example.smartshopping.domain.product.registration.ProductRegistrationRequest
-import com.example.smartshopping.domain.product.registration.ProductRegistrationService
-import com.example.smartshopping.domain.product.toProductListItemResponse
-import com.example.smartshopping.domain.product.toProductResponse
-import com.example.smartshopping.domain.review.ReviewRequest
-import com.example.smartshopping.domain.review.ReviewService
+import com.example.smartshopping.domain.request.ProductRegistrationRequest
+import com.example.smartshopping.domain.request.ReviewRequest
+import com.example.smartshopping.domain.response.toProductListItemResponse
+import com.example.smartshopping.domain.response.toProductResponse
+import com.example.smartshopping.domain.service.ProductImageService
+import com.example.smartshopping.domain.service.ProductRegistrationService
+import com.example.smartshopping.domain.service.ProductService
+import com.example.smartshopping.domain.service.ReviewService
 import com.example.smartshopping.entity.Product
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.*
