+++ src/main/kotlin/com/example/smartshopping/presentation/controller/PurchaseApiController.kt
@@ -1,9 +1,9 @@
 package com.example.smartshopping.presentation.controller
 
 import com.example.smartshopping.common.ApiResponse
-import com.example.smartshopping.domain.purchase.PurchaseRequest
-import com.example.smartshopping.domain.purchase.PurchaseService
-import com.example.smartshopping.domain.purchase.toPurchaseResponse
+import com.example.smartshopping.domain.request.PurchaseRequest
+import com.example.smartshopping.domain.response.toPurchaseResponse
+import com.example.smartshopping.domain.service.PurchaseService
 import com.example.smartshopping.entity.PurchaseRecord
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.*
